[//gpu] Set default value for Mac texture target in SharedImageCaps
I had hoped to initialize this state to 0 and enforce that clients would
appropriately initialize it when creating SICaps. However, unittests
broadly create and use SharedImageCapabilities instances without
initializing this field. The lack of initialization causes
ClientSharedImage creation to crash in the context of these tests when
it uses this state in setting its texture target.

These tests currently work because legacy global variable that stores
the Mac texture target is initialized to GL_TEXTURE_RECTANGLE_ARB, which
is a valid value for ClientSharedImage to use. In the interest of moving
along the production change, this CL changes SharedImageCapabilities to
also initialize this state to GL_TEXTURE_RECTANGLE_ARB. In the fullness
of time, I still want to fix all of the unittest creation sites and
make the change described above. However, there is no reason for that to
block changing the production code away from using the legacy global
variable.

Bug: 41494843
Change-Id: I2e63696d9640413fa45fb0ed7250abfc3c34b024
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5401809
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280731}