[editing] createLink command dispatches an input event with data and inputType which is set as `insertLink`
According to the spec:
https://w3c.github.io/input-events/#interface-InputEvent-Attributes
When inserting a link, `inputType` value should be set to `insertLink`.
The `InputType` enumeration type, which is defined in
`third_party/blink/renderer/core/events/input_event.h`
is not in compliance with the spec and lacks the `inesetLink` value.

The `createLink` test case, which is defined in
`external/wpt/input-events/input-events-exec-command.html`,
will be failed due to the absence of `data` and `inputType`
attributes in the event. These test case only failed in Chrome,
while Firefox and Safari were ok.
`createLink` should dispatch an input event with `data`
and `inputType` which is set as `insertLink`.

Bug: 331836632
Change-Id: I7e789bac0a32713bbf1f0b27832d93c9ca1833d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5392761
Reviewed-by: Kent Tamura <tkent@chromium.org>
Auto-Submit: Peng Zhou <zhoupeng.1996@bytedance.com>
Commit-Queue: Peng Zhou <zhoupeng.1996@bytedance.com>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280738}