PDF OCR: Remove trailing whitespace in tree dump for tests
OCR is subject to detect whitespace after a sentence and put it in its
results, which is quite difficult for us to anticipate when. We recently
observed this behavior in PDFOCRIntegrationTest after upreving the
screen-ai library to version 123.6 [1]. To mitigate this unexpected
behavior, remove trailing whitespace in OCR results inserted into a
static text node for tests using `DumpPdfAccessibilityTree`.

[1] https://crrev.com/c/5388533

AX-Relnotes: n/a.
Bug: 331659257
Change-Id: I6a6a828be65710d23555452cc5dcf7913ffd400f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5401607
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Kyungjun Lee <kyungjunlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1280752}