Remove nigi@chromium.org from pdf/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/nigi@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: If416ac965c9d6ab423503b3518b8bae304ea3ed8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5155832
Commit-Queue: Lei Zhang <thestig@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280754}