utr: Fix markdown printer crash when recipe json is empty
As-is, the utr will crash with the following error when the recipe
passes:
TypeError: Input data should be a string, not <class 'NoneType'>

That's just because we were returning None in _run_recipe() when we
should've been returning a string.

This fixes that + adds a test case that would've caught it.

Bug: 41492688
Change-Id: Ibe677039117cd2902e6ddae3dd1249037cd6d8bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5410831
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280828}