[ TabRestore ] Store saved id of groups in Tab and Group Objects
Stores the saved id of a group when it is closed so we can restore the saved version later when TabRestore is triggered.

Moves the logic to disconnect saved groups into the BrowserTabStripModelDelegate so other callers do not need to manually do this themselves.

Change-Id: I2eda70a1b8403380da32ec45b6cd70b5f8f70e77
Bug: 330769406
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5385301
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Darryl James <dljames@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280831}