[Protected Audience] Update DoGetInterestGroupsForOwner to load data in one pass
Currently, we query the interest group names first, and then for each
name, we query for matching entries. We switch to load all data at once,
and do the sorting in C++. Offline testings have shown that this could
reduce the load time by 30% on a large database.

Also, fix/improve a test. We modify the expiry time to make the
ordering deterministic.

Bug: 331902978
Change-Id: I6162cba36ef0b6f9693e5377d5ac2b1bc93e565e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5403172
Reviewed-by: Russ Hamilton <behamilton@google.com>
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280848}