Add keep sorted lint to vector icons build file
To test it I tried moving one of the items to be unsorted and when I
tried uploading a CL it said I needed to run `git cl format` which
moved the item back to the correct sorted order.

The go/keep-sorted analyzer is available through AyeAye and is used
heavily internally. Hopefully this can help simplify workflows in
Chromium now too.

Change-Id: I1177f15ec583bf872afe8225237fad056c31d037
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5410183
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Alison Gale <agale@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280883}