Tab Organization: Fix GetDisplayName crash
This crash was caused by prematurely removing a session containing
multiple groups after accepting the first group, in the single
organization UI. The previous logic referenced session->IsComplete(),
which returns false if there are remaining groups to accept. This now
looks directly at the request state, as the presence of remaining groups
is irrelevant to this part of the code.

Bug: 332077792
Change-Id: If55b37c45fc3bafc9f6bcddbb150fbd54b971aa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5411039
Commit-Queue: Emily Shack <emshack@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280885}