Lacros-Test: Enable AppSessionRestoreTest.RestoreAppMinimized.
On Lacros, minimizing the browser window runs asynchronously
due to the asynchronous wayland UI flow, therefore, the test
code needs to wait for the expected window being minimized,
instead of assuming it happens synchronously. This cl adjust
the test code to wait for the window to be minimized so that
the test is enabled for Lacros.

Bug: 40794915
Change-Id: Iac66f3bac0f58b8732888559f338474db7ba0069
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5407836
Commit-Queue: Jenny Zhang <jennyz@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280906}