Port AppSessionRestoreTest.CtrlShiftTRestoresAppsCorrectly to Lacros.
This test fails for Lacros in browser_tests since the testing web
app is not published to app service, and session restore service
queries app service to decide whether to restore the window as an
app window. Web Apps replies on app service crosapi to publish the
app, crosapi is not enabled in browser_tests, therefore, moving
this test to lacros_chrome_browsertests in which crosapi is enabled
and version skew tests are in place.

Bug: 40794915
Change-Id: I40ef918829725c39bf7efc30ac6e3bd56e95940c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5408274
Commit-Queue: Jenny Zhang <jennyz@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280908}