[Cast Streaming] Remove legacy frame sender.
This removes the legacy frame sender used to bridge the streaming
sender API to the pre-libcast Cast sender.  This will allow a large
portion of the legacy Cast transport to be removed in a follow-up CL.

This required updating several unit tests which were never ported
to use libcast.  Either they are updated to use libcast objects, or
test cases are deleted if they test functionality that is internal
to libcast.

Bug: b/41481348
Fixed: b/40226049
Change-Id: I2ecbda8295ca5dcf9cce818a6e7139a1434e22a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5173522
Commit-Queue: Mark Foltz <mfoltz@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280912}