[Extensions] Relax service worker script checks
When evaluating a service worker script for an extension in the
renderer, we (immediately) execute a script to set up a couple of the
extension APIs. We expect this to always succeed because:
* The script is included with Chrome
* It's executed immediately when the service worker context starts
  (i.e., before any untrusted code runs).

However, this unfortunately isn't _always_ guaranteed to work. If the
chrome resources are (even slightly) corrupted, this could throw a
runtime error, or other pieces could potentially influence the execution
of the script. This doesn't appear to be related to any particular
extension or client.

Relax these checks and gracefully fail instead.

Bug: 41486930
Change-Id: I60feb379d9111d662404390a9c5d27fbb5120ce9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5411235
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280933}