focus-mode: Show media action icons and equalizer animation
This CL will show the icons for a single playlist view. When the
playlist isn't playing, we show the play icon. When the playlist is
playing, if a mouse hovers on the thumbnail of the playlist, we show the
stop icon; otherwise, we show the euqalizer animation instead. Also, we
will show a check icon for the playing playlist. We don't add the logic
to stop other playlists when playing a new playlist in this CL.

Bug: b/325346328
Test: Manual
Change-Id: I345a9af04a3d181f065a1fd00b52863b379549c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5398047
Reviewed-by: Richard Chui <richui@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Hongyu Long <hongyulong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280944}