Reland: Detect AAC isn't supported on Windows N.
The previous CL:https://chromium-review.googlesource.com/c/chromium/src/+/5379757
is reverted in the failure of Win11 ARM64, where kPlatformAudioEncoder isn't
enabled so the failure is correct.

The change is same for every places except the test file
media_recorder_handler_unittest.cc in that it detects
AAC is supported before running test and merge IsAacCodecInUnSupportedPlatform
function to the IsTargetAudioCodecSupported.

---------------------------------------------------------------------
Description of the previous CL (5379757).

In the context of Windows N SKU, the isTypeSupported function
does not include support for AAC (Advanced Audio Coding) encoding.
In standard Windows operating systems, AAC encoding relies on the
Media Foundation library. However, this library is not installed
by default in Windows N SKU, resulting in the lack of AAC encoding support.

The straightforward solution is to check whether the system is
running Windows N SKU and return false for AAC support.

An alternative approach involves examining the presence of
the Media Foundation binaries (such as mf.dll). Unfortunately,
this file existence check cannot be performed directly in the
renderer process. The GPU process already handles this check,
but the isTypeSupported API operates synchronously and cannot
utilize a Mojo-based solution.

It’s worth noting that Windows N SKU is optional for European
countries and Korea, although it is not widely used. Therefore,
a simple check based on the product type could be a more effective solution.

Bug: 329698355
Change-Id: Icce607952988d6ea96d9ecf5b0ab0ad63dfbe702
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5407269
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Sunggook Chue <sunggch@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1280949}