Attempt to fix dangling raw_ptr in ResourceLoadScheduler
test_task_runner and its mock clock can be destroyed after the test
ends and before we cleaning up the testing env, and the scheduler
is destroyed while we're cleaning up the test env, so it is possible
that when destroying scheduler, it clock_ has been deleted, which
triggers dangling raw_ptr failure.
This CL attempts to fix the flakiness by resetting the scheduler's
clock before finishing the test.

Fixed: 328504978
Change-Id: I0830138a9f6bd91046d98942045795e4cb0b9f6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5407130
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280999}