[Test Automation] ScreenshotSurface convenience verb
Adds a test verb that takes a screenshot of the entire surface (i.e.
Widget) associated with a given element.

This is especially useful for taking screenshots of dialogs, where e.g.
the footer of the menu is not part of the [Bubble]DialogView.

Change-Id: Ia4ea09a853903afabfc5d6dc2e780282269d1c5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5406870
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1281005}