media/gpu/V4L2StatefulVD: Support Decode() after Reset()
V4L2StatefulVD, like other media::VideoDecoder implementations,
are normally called in a sequence like:
- Initialize - Decode - Decode ... - Reset - Initialize - Decode ...

VDVideoDecodeAccelerator does something else: it skips the
Initialize after Reset and expects the next Decode to work:

- Initialize - Decode - Decode ... - Reset - Decode ...

media::VideoDecoder interface does not specify one or the other
as being correct or incorrect [1], so while strange, VDVDA approach
seems to be OK.

On top of that, said VDVDA tends to produce the Reset() call when
the media::VideoDecoder impl (in this case, V4L2StatefulVD) is half
way thru its own initialization (concretely of the CAPTURE queue).
This happens concretely when V4L2StatefulVD asks its owner to
PickDecoderOutputFormat(): this is a simple synchronous call for
Chrome clients, but a process hop for ARCVM clients. This situation
is so specific that there's a unit test for it [2]

ToT V4l2VideoDecoder/BackendStateful solve this situation by
separating  its internal initialization into multiple steps
(ChangeResolution - ContinueChangeResolution etc), so that it can
pick up where it left in the first Decode call after Reset [3]

This CL takes a simpler approach: it props up V4L2StatefulVD::Reset()
to nuke both OUTPUT and CAPTURE queues and allows for a Decode()
call to find the queues not configured, calling Initialize() in
that case.

[1] https://source.chromium.org/chromium/chromium/src/+/main:media/base/video_decoder.h;l=49-107;drc=9f7e05d16d893bc9c542f027a2be74fba1773aa4
[2] https://source.chromium.org/chromium/chromium/src/+/main:media/gpu/test/video_decode_accelerator_tests.cc;l=572-594;drc=60a9cfbee3fb02198de22f06aa518be7f4aee40e
[3] https://source.chromium.org/chromium/chromium/src/+/main:media/gpu/v4l2/v4l2_video_decoder_backend_stateful.cc;l=161-163;drc=83786cefc38b3e9cc29275b975ed091bf706a7ae


Bug: 328002324
Change-Id: I3583f113493a6069ec28ad23a72acd3bf2fa3003
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5398807
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1281017}