Files app: Change DirectoryModel to allow FileKey as current directory
As a preparation for the materialized views, which aren't backed by an
entry, allow the current directory to be a FileKey.

Add DirectoryContents.getFileKey(), DiretoryModel.getCurrentFileKey()
and DirectoryModel.getCurrentFileData() as the counter part of getDirectoryEntry() and getCurrentDirEntry().

Update the event `directory-changed` to include the previous and the
current key and make the entries (previous and current) optional.

Bug: b:327532840, b:327534678
Change-Id: I13213b05f5499a369f704ea768e08525c9b327ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5404306
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Wenbo Jie <wenbojie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1281021}