Reland "Add views for the PaymentsWindowUserConsentDialog"
This is a reland of commit 2a91add64b7e9531dfe06edb9c9520df72faecd6

The original test failure was because we did not wait for the dialog
to show after calling TriggerDialog(), and instantly tried to call
PressButton(). Dialog showing is asynchronous, so this resulted in a
flaky test because there can be times where we attempt to press a
button before the dialog has actually shown. Thus when we try to wait
after pressing the button, it will time out because the dialog will
stay shown, and the test will fail.

To fix this, a WaitForShow() call was added after every
TriggerDialog() call to ensure the dialog is fully shown before
proceeding to the next step.

Original change's description:
> Add views for the PaymentsWindowUserConsentDialog
>
> This CL adds the view implementation and creation for the
> PaymentsWindowUserConsentDialog, as well as its corresponding browser
> tests. This dialog will be triggered to get user consent for initiating
> a payments window flow, for which we currently only have VCN 3DS.
>
> Mock: https://screenshot.googleplex.com/3CBZNUJ7hvNyBn9
>
> Implementation: https://screenshot.googleplex.com/3qFCx7yRXKGJbRy
>
> Bug: 1517762
> Change-Id: I1fa422faa33ce68f257973347def9cce3f456d2b
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5338589
> Commit-Queue: Vinny Persky <vinnypersky@google.com>
> Reviewed-by: Siyu An <siyua@chromium.org>
> Reviewed-by: Dana Fried <dfried@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1280290}

Bug: 1517762
Change-Id: Ia80e91c293d3ea72463abaaa05d04588274ca965
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5408223
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Cr-Commit-Position: refs/heads/main@{#1281041}