[iOS] Change the Grid's move to take items instead of index
This allows it to move groups via drag and drop.
It is also removing the grid command to move item as it wasn't
necessary: it was called in response to a move event on the
WebStateList, so it was a no-op.

Based on: https://crrev.com/c/5405944

Bug: 331749274
Change-Id: Ib1d67935628416db01508124de77b850e5faf556
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5402585
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1281113}